lesson_10/requirements.txt
pytest==9.0.2
selenium==4.39.0
allure-pytest==2.15.3